rename handlePayment to execute
