+++ extension/src/paymentHandler/get-origin-keys.handler.js
@@ -1,10 +1,10 @@
 const pU = require('./payment-utils')
 const c = require('../config/constants')
-const { getOriginKeys } = require('./web-component-service')
+const { getOriginKeys } = require('../web-component-service')
 
-async function handlePayment (paymentObject) {
-  const getOriginKeysRequest = paymentObject.custom.fields.getOriginKeysRequest
-  const { request, response } = await getOriginKeys(getOriginKeysRequest)
+async function execute (paymentObject) {
+  const getOriginKeysRequestObj = JSON.parse(paymentObject.custom.fields.getOriginKeysRequest)
+  const { request, response } = await getOriginKeys(getOriginKeysRequestObj)
   return {
     actions: [
       pU.createAddInterfaceInteractionAction({
@@ -15,4 +15,4 @@ async function handlePayment (paymentObject) {
   }
 }
 
-module.exports = { handlePayment }
+module.exports = { execute }
