+++ extension/src/paymentHandler/get-payment-methods.handler.js
@@ -1,10 +1,10 @@
 const pU = require('./payment-utils')
 const c = require('../config/constants')
-const { getPaymentMethods } = require('./web-component-service')
+const { getPaymentMethods } = require('../web-component-service')
 
-async function handlePayment (paymentObject) {
-  const getPaymentMethodsRequest = paymentObject.custom.fields.getPaymentMethodsRequest
-  const { request, response } = await getPaymentMethods(getPaymentMethodsRequest)
+async function execute (paymentObject) {
+  const getPaymentMethodsRequestObj = JSON.parse(paymentObject.custom.fields.getPaymentMethodsRequest)
+  const { request, response } = await getPaymentMethods(getPaymentMethodsRequestObj)
   return {
     actions: [
       pU.createAddInterfaceInteractionAction({
@@ -15,4 +15,4 @@ async function handlePayment (paymentObject) {
   }
 }
 
-module.exports = { handlePayment }
+module.exports = { execute }
